refactor: make changes based on usage
