+++ package.json
@@ -1,24 +1,34 @@
 {
-  "private": true,
-  "name": "@drashland/drash",
-  "version": "3.0.0-preview.202308281922",
+  "name": "@crookse/drash",
+  "version": "3.0.0-beta.202310151150",
   "description": "A microframework for building JavaScript HTTP applications",
   "license": "GPL-3.0",
-  "repository": "git@github.com:drashland/drash.git",
+  "repository": {
+    "type": "git",
+    "url": "git+ssh://git@github.com/drashland/drash.git"
+  },
   "author": "Drash Land",
   "contributors": [
     {
-      "name": "Breno Salles"
+      "name": "Breno",
+      "url": "https://github.com/Guergeiro"
     },
     {
-      "name": "Edward Bebbington"
+      "name": "Edward",
+      "url": "https://github.com/ebebbington"
     },
     {
-      "name": "Eric Crooks"
+      "name": "Eric",
+      "url": "https://github.com/crookse"
     },
     {
-      "name": "Sara Gee"
+      "name": "Sara",
+      "url": "https://github.com/saragee3"
     }
+  ],
+  "files": [
+    "./lib/cjs",
+    "./lib/esm"
   ],
   "devDependencies": {
     "@babel/preset-env": "^7.22.10",
