--- package.lib.json
@@ -1,23 +1,0 @@
-{
-  "private": true,
-  "name": "@drashland/drash",
-  "version": "3.0.0-preview.202308281922",
-  "description": "A microframework for building JavaScript HTTP applications",
-  "license": "GPL-3.0",
-  "repository": "git@github.com:drashland/drash.git",
-  "author": "Drash Land",
-  "contributors": [
-    {
-      "name": "Breno Salles"
-    },
-    {
-      "name": "Edward Bebbington"
-    },
-    {
-      "name": "Eric Crooks"
-    },
-    {
-      "name": "Sara Gee"
-    }
-  ]
-}
