+++ scripts/lib_builder/build_intermediary
@@ -6,6 +6,6 @@
   deno fmt \
   && deno run -A ./scripts/lib_builder/build_esm_lib.ts \
     --debug \
-    --workspace=./.drashland/lib/intermediary \
+    --workspace=./lib/intermediary \
     --copy-files=./src
 )
