+++ tests/compat/bun/v0.x/modules/RequestChain/native/default-behavior/app.ts
@@ -19,11 +19,11 @@
  * Drash. If not, see <https://www.gnu.org/licenses/>.
  */
 
-import { HTTPError } from "../../../../../../../../.drashland/lib/esm/core/errors/HTTPError";
-import { Status } from "../../../../../../../../.drashland/lib/esm/core/http/response/Status";
-import { StatusCode } from "../../../../../../../../.drashland/lib/esm/core/http/response/StatusCode";
-import { StatusDescription } from "../../../../../../../../.drashland/lib/esm/core/http/response/StatusDescription";
-import * as Chain from "../../../../../../../../.drashland/lib/esm/modules/RequestChain/mod.native";
+import { HTTPError } from "../../../../../../../../lib/esm/core/errors/HTTPError";
+import { Status } from "../../../../../../../../lib/esm/core/http/response/Status";
+import { StatusCode } from "../../../../../../../../lib/esm/core/http/response/StatusCode";
+import { StatusDescription } from "../../../../../../../../lib/esm/core/http/response/StatusDescription";
+import * as Chain from "../../../../../../../../lib/esm/modules/chains/RequestChain/mod.native";
 
 export const protocol = "http";
 export const hostname = "localhost";
