+++ tests/compat/deno/v1.x/modules/RequestChain/native/request-in-context-object/app.ts
@@ -21,7 +21,7 @@
 
 import { StatusCode } from "../../../../../../../../src/core/http/response/StatusCode.ts";
 import { StatusDescription } from "../../../../../../../../src/core/http/response/StatusDescription.ts";
-import * as Chain from "../../../../../../../../src/modules/RequestChain/mod.native.ts";
+import * as Chain from "../../../../../../../../src/modules/chains/RequestChain/mod.native.ts";
 import { Status } from "../../../../../../../../src/core/http/response/Status.ts";
 
 export const protocol = "http";
