+++ tests/compat/deno/v1.x/modules/RequestChain/native/resource-groups/app.ts
@@ -20,7 +20,7 @@
  */
 
 import { HTTPError } from "../../../../../../../../src/core/errors/HTTPError.ts";
-import * as Chain from "../../../../../../../../src/modules/RequestChain/mod.native.ts";
+import * as Chain from "../../../../../../../../src/modules/chains/RequestChain/mod.native.ts";
 import { StatusCode } from "../../../../../../../../src/core/http/response/StatusCode.ts";
 import { StatusDescription } from "../../../../../../../../src/core/http/response/StatusDescription.ts";
 import { ResourceGroup } from "../../../../../../../../src/standard/http/ResourceGroup.ts";
