+++ tests/compat/deno/v1.x/modules/RequestChain/polyfill/concurrency/app.ts
@@ -22,7 +22,7 @@
 import { Status } from "../../../../../../../../src/core/http/response/Status.ts";
 import { StatusCode } from "../../../../../../../../src/core/http/response/StatusCode.ts";
 import { StatusDescription } from "../../../../../../../../src/core/http/response/StatusDescription.ts";
-import * as Chain from "../../../../../../../../src/modules/RequestChain/mod.polyfill.ts";
+import * as Chain from "../../../../../../../../src/modules/chains/RequestChain/mod.polyfill.ts";
 
 export const protocol = "http";
 export const hostname = "localhost";
