+++ tests/compat/node/node-v18.x/modules/RequestChain/polyfill/node-http-converted-to-web-api/app.ts
@@ -21,11 +21,11 @@
 
 import { IncomingMessage, ServerResponse } from "node:http";
 
-import { HTTPError } from "../../../../../../../../.drashland/lib/esm/core/errors/HTTPError";
-import { StatusCode } from "../../../../../../../../.drashland/lib/esm/core/http/response/StatusCode";
-import { StatusDescription } from "../../../../../../../../.drashland/lib/esm/core/http/response/StatusDescription";
-import * as Chain from "../../../../../../../../.drashland/lib/esm/modules/RequestChain/mod.polyfill";
-import { Status } from "../../../../../../../../.drashland/lib/esm/core/http/response/Status";
+import { HTTPError } from "../../../../../../../../lib/esm/core/errors/HTTPError";
+import { StatusCode } from "../../../../../../../../lib/esm/core/http/response/StatusCode";
+import { StatusDescription } from "../../../../../../../../lib/esm/core/http/response/StatusDescription";
+import * as Chain from "../../../../../../../../lib/esm/modules/chains/RequestChain/mod.polyfill";
+import { Status } from "../../../../../../../../lib/esm/core/http/response/Status";
 
 export const protocol = "http";
 export const hostname = "localhost";
