+++ tests/middleware/deno/v1.x/modules/middleware/ETag/mod_test.ts
@@ -30,7 +30,7 @@ import {
 import { StatusCode } from "../../../../../../../src/core/http/response/StatusCode.ts";
 import { StatusDescription } from "../../../../../../../src/core/http/response/StatusDescription.ts";
 import { Method } from "../../../../../../../src/core/http/request/Method.ts";
-import * as Chain from "../../../../../../../src/modules/RequestChain/mod.native.ts";
+import * as Chain from "../../../../../../../src/modules/chains/RequestChain/mod.native.ts";
 import {
   defaultOptions,
   ETag,
