+++ tsconfig.build.esm.json
@@ -1,8 +1,8 @@
 {
   "extends": "./tsconfig.build.base.json",
   "compilerOptions": {
-    "outDir": "./.drashland/lib/esm",
-    "rootDir": "./.drashland/lib/intermediary"
+    "outDir": "./lib/esm",
+    "rootDir": "./lib/intermediary"
   },
-  "include": ["./.drashland/lib/intermediary"]
+  "include": ["./lib/intermediary"]
 }
